+++ src/Sockets.php
@@ -10,24 +10,63 @@ use Innmind\IO\{
 };
 use Innmind\Immutable\Attempt;
 
-interface Sockets
+final class Sockets
 {
+    private Config $config;
+
+    private function __construct(Config $config)
+    {
+        $this->config = $config;
+    }
+
+    /**
+     * @internal
+     */
+    public static function of(Config $config): self
+    {
+        return new self($config);
+    }
+
     /**
      * This method will fail if the socket already exist
      *
      * @return Attempt<Server>
      */
-    public function open(Address $address): Attempt;
+    public function open(Address $address): Attempt
+    {
+        return $this
+            ->config
+            ->io()
+            ->sockets()
+            ->servers()
+            ->unix($address);
+    }
 
     /**
      * This will take control of the socket if it already exist (use carefully)
      *
      * @return Attempt<Server>
      */
-    public function takeOver(Address $address): Attempt;
+    public function takeOver(Address $address): Attempt
+    {
+        return $this
+            ->config
+            ->io()
+            ->sockets()
+            ->servers()
+            ->takeOver($address);
+    }
 
     /**
      * @return Attempt<Client>
      */
-    public function connectTo(Address $address): Attempt;
+    public function connectTo(Address $address): Attempt
+    {
+        return $this
+            ->config
+            ->io()
+            ->sockets()
+            ->clients()
+            ->unix($address);
+    }
 }